requirements.txt
google-cloud==0.34.0
google-cloud-pubsub==0.40.0
google-cloud-error-reporting==0.30.1
requests==2.21.0